# src/docs/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Evently API",
    "description": "API for Evently",
    "version": "1.0.0"
  },
  "host": [
    "theodev.myftp.org:89"
  ],
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter your Bearer token in the format **Bearer <token>**"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/events": {
      "post": {
        "tags": [
          "Event"
        ],
        "description": "Create new event. Optionnellement, ajoutez des amis via leur username (tableau \"friends\").",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Event name"
                },
                "description": {
                  "type": "string",
                  "example": "Event description"
                },
                "folderID": {
                  "type": "boolean",
                  "example": false
                },
                "recurrencePattern": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "example": "string"
                    },
                    "enum": {
                      "type": "array",
                      "example": [
                        "daily",
                        "weekly",
                        "monthly",
                        "yearly"
                      ],
                      "items": {
                        "type": "string"
                      }
                    },
                    "example": {
                      "type": "string",
                      "example": "daily"
                    }
                  }
                },
                "targetDate": {
                  "type": "string",
                  "example": "2024-01-01T00:00:00.000Z"
                },
                "categoryID": {
                  "type": "number",
                  "example": 1
                },
                "friends": {
                  "type": "array",
                  "example": [
                    "username1",
                    "username2"
                  ],
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      },
      "get": {
        "tags": [
          "Event"
        ],
        "description": "Get user events",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/events/{id}": {
      "put": {
        "tags": [
          "Event"
        ],
        "description": "Update event. Optionnellement, ajoutez des amis via leur username (tableau \"friends\").",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Updated event name"
                },
                "description": {
                  "type": "string",
                  "example": "Updated event description"
                },
                "recurrencePattern": {
                  "type": "string",
                  "example": "weekly"
                },
                "interval": {
                  "type": "number",
                  "example": 2
                },
                "targetDate": {
                  "type": "string",
                  "example": "2024-02-01T00:00:00.000Z"
                },
                "categoryID": {
                  "type": "number",
                  "example": 2
                },
                "friends": {
                  "type": "array",
                  "example": [
                    "username1",
                    "username2"
                  ],
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      },
      "delete": {
        "tags": [
          "Event"
        ],
        "description": "Delete event",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/events/share/{id}": {
      "post": {
        "tags": [
          "Event"
        ],
        "description": "Generate an invite token for an event (only for the event creator)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/events/join": {
      "post": {
        "tags": [
          "Event"
        ],
        "description": "Rejoindre un event à l\\'aide d\\'un token d\\'invitation.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "inviteToken": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event rejoint avec succès.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Event rejoint avec succès"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "L'utilisateur est déjà le créateur ou déjà membre de l'event.",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 400
                },
                "error": {
                  "type": "string",
                  "example": "Vous êtes déjà le propriétaire de cet event"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Event introuvable pour le token fourni.",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 404
                },
                "error": {
                  "type": "string",
                  "example": "Event not found for the given token"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/events/leave/{id}": {
      "post": {
        "tags": [
          "Event"
        ],
        "description": "Quitter un event (si l\\'utilisateur n\\'en est pas le créateur).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID de l'event à quitter"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "Event quitté avec succès."
          },
          "400": {
            "description": "Le créateur ne peut pas quitter son propre event.",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 400
                },
                "error": {
                  "type": "string",
                  "example": "Le propriétaire ne peut pas quitter cet event"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Event introuvable ou l'utilisateur n'est pas membre.",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 404
                },
                "error": {
                  "type": "string",
                  "example": "Event not found or you are not a member"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/events/invitation": {
      "get": {
        "tags": [
          "Event"
        ],
        "description": "Récupérer les invitations à traiter pour l’utilisateur (status \"invited\" ou \"pending\").",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/events/invitation/{eventId}": {
      "put": {
        "tags": [
          "Event"
        ],
        "description": "Mettre à jour le status d\\'une invitation pour un event. Fournir un status : \"accepted\", \"pending\" ou \"refused\" (le refus mène à la suppression de l\\'invitation).",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
...
  },
...
  "security": [
    {
      "Bearer": []
    }
  ]
}